This submission corrects Trac Ticket #39.
A memory leak was discovered in the situation where a WKT definition is imported and the NameMapper fails to produce a datum match between the definition and the CS-MAP dictionaries.  The leak consist of a single copy (per processed WKT definitioin which produces this relatively rare situation) of the cs-Dtdef_ structure, about 200 bytes.

Ref LUC009NTO01.txt