CS-MAP Trac Ticket # 109.
When performing an inverse datum shift using the NADCON iterative calculation, points on the very edge of a NADCON grid file can produce an iterative sequence which doesn't converge, and is therefore reported as a fatal error.  This occurs as the iterative approximation algorithm alternates between being with the coverage of the file and outside the coverage of the file.  Thus, a successful convergence is not achieved, and more importantly, the condition is reported as a fatal condition.

Technically, in the NADCON environment, a point exactly on the north or eastern edge of the grid file extents is outside the coverage. So, that is how we fixed this.  A point within a very very small distance of either the northern or eastern edge of the NADCON grid file extents is now considered to be outside the extents.  This substantially eliminates this problem.