Fix Trac Ticket # 9 -- Changed the Geodetic Transformation dictionary entry for the German BeTA2007 datum transformation grid shift data file from BeTA2007.gsb to BETA2007.gsb.  Obviously, this change in case only affects operation on a Linux system.
Note that the documentation of this datum shift data file uses the name BeTA 2007 to refer to the datum shift.  However, if you download the file form the official site, you get a file named BETA2007.gsb.  So, I guess BETA2007.gsb is the name we should use for the file.