RFC-8, Phase Two -- Submitting the merge of the trunk with this sandbox in preparation of merging this sandbox back into the trunk.  I confess to being rather confused in maintaining this sandbox.  In maintaining the sandbox and the trunk simultaneously, working with two different clients, on several occasions found myself making edits in the wrong location.  I.e. using Visual Studio which was set to activate the last open solution, I erroneously was editing the trunk when I should have been editing the sandbox, and vice versa. I have disabled this feature of Visual Studio, and believe this submission corrects all of the erroneous edits.
If testing shows that all of these discrepancies have been corrected, this will be the final submission to this sandbox prior to a "reintegrate" merge back into the trunk.  Generally, this submission adds the "Console Utility" code used to generate the 48hpgn.l?s set of data files, and a bunch of RFC-8 clean up edits which got into the trunk which really belonged in the sandbox.

It also includes some Trac ticket fixes which were correctly to the trunk, and needed to be synch'ed with the sandbox before the final merge.