Working related Tickets: #146, #155, and some rework on 151.
Fixing 146 required fixing a bug in the CS_extractDbl function in the CS_general.c module.  Clearly a coding error which worked for the initial case (WW15MGH.GRD), but not the general case.  The user supplied patch could not be used as it was out of date.

The problem reported in ticket 155 was real, but only in a Unix/Linux environment.  In Windows, one cannot arrange to use security to prevent the creation of files in a directory which one otherwise has access to.  At least I couldn't figure out a way to do it.  The problem, it turns out is not limited to EGM96; but also applies to several other modules which are programmed to "pre-process" large text files into "temporary" binary versions for performance purposes.  These modules are OSTN97, OSTN02, OSGM91, and the Japanese GDK2000 module which processes the .par files.

The solution to the write protected directory problem was to recognize a new environmental variable "CS_MAP_SIR_RW".  All five modules have been modified such that should the standard Dictionary directory be write protected, the value of the CS_MAP_DIR_RW environment variable will be used as the location for the "temporary" binary file (i.e. binary shadow file). Iterative environmental variable substitution is performed prior to the value of this variable being used.

The five modules involved here were also adjusted so that compilation of the ASCII text files will be performed when appropriate. The feature introduced by ticket #151 is preserved.  That is, presence of the shadow binary file is accepted even in the absence of the primary ASCII text file.  However, the ability to compile an ASCII text file was restored.  This is appropriate in the Linux environment where the byte order of the platform may not be little endian.  Also, restoring the compilation capability is necessary for users to add new and different geoid height data files to the systems repertoire.

Finally, as byte order problems are possible, the constructors of four of the above modules were modified to include a test case to insure that there is no byte order issue with the resulting binary file (regardless of compilation or not) before a successful object construction is returned.  These functions existed for four of the objects.  Implementation of a similar test for the Japanese .par object is left as a TO DO item.  Clearly, Test 4 of the Console Test Utility includes some substantial testing of the .par object sufficient for this submission.  A new error message was implemented: cs_SELF_TEST which is emitted in the case of the self test feature failing in the construction of these objects.

Testing was performed using the Console Test Utility, with special test cases added to the special case CStestT.c.