Trac Ticket #198 -- Have added 184 coordinate systems, representing the 92 zones of the Indiana DOT GCS systems.  Each zone has two distinct CRS definitions: one in meters and the second in US Survey Feet.   Appropriate entries has also been made in the category file, the NameMapper file, and the TEST.DAST file.  See  ​http://www.in.gov/indot/3397.htm for detailed information.  In the Data sub-directory, a file named InGCS.csv contains all the particulars of the 92 zones involved in this update.
All changes were made programmatically to reduce the chance of a typographical error.  In so doing, files such as coordsys.asc and NameMapper.csv were read into memory edited programmatically, and then written back out.  IN so doing, some extraneous white space has been deleted in coordsys,asc, and the order of some of the entries in the NameMapper have changed.  Order of entries in the NameMapper file have no affect on the operation of the NameMapper.

A test for each added system has been added to both the TEST.DAT file and the TEST_All.DAT file.

One new module now exists in the CsMapConsoleUtilities project.  Note that the code in the CsMapConsoleUtilities project is only used during maintenance operations.  There is no need for this project to get built (neither release or debug) unless you are going to use one of the utilities for maintenance purposes.  You used to be able to mark a project as "do not build", but that seems to have disappeared. If you know where it went, please disable the CsMapConsoleUtilities from the solution; it will save some build time.