This submission includes corrections for Nova Scotia and Canadian conversions and transforms for defects discovered during testing.  The submission includes test data, BUT NOT the grid shift files required to do these tests.
In addition to the Canadian and French grid shift data files which had to obtained separately, add two more Canadian shift data files:

NS7783021.GSB    -- Nova Scotia specific ATS77 to NAD83.2010 (CSRS.2010)
CVG69710.GSB     -- CSRS (NAD83.1997) to CSRS.2010 for all of Canada

References to these files in the Geodetic Transformation dictionary have been commented out.  These references will need to restored, and the dictionaries re-compiled, in order to test the new systems which depend upon them.

The test cases for these new systems have been added to the TEST_All.DAT file in the TestCpp directory.  You need to explicitly provide the name of this file on the command line for the Console Test module to use it.  (The default is TEST.DAT, which does not include tests which depend upon grid shift data files which cannot be distributed as part of an open source project.)

Source of the test data is documented in the test files.