Working Trac Ticket #207 -- Improved the Geodetic Path checker to verify that the chain of datums from the first entry through the last entry are a continuous chain, and that the source datum and target datum match the ends of the transformation chain.  Add the use of this improved checker to the three places where Geodetic Paths are fetched and used.  Thus, users will get an appropriate error message in the future at run time. Code in CS_testT.cpp was used to test the fix; you need to queer the geodetic path definition named NSRS11_to_WGS84 and compile the queered result for this code to make the appropriate test.
This check should be added to the dictionary compiler, but the environment there is a little different, so we'll leave that to the effort to fix Trac Ticket #208.

Two error conditions/messages were added to cs_map.h and CS_error.c.  Since I was submitting a change to cs_map.h, I took the liberty of correcting a whole flock of spelling errors.  The changes significant to this fix are four or five lines starting about 658 in cs_map.h and 847 in cs_error.c.