Added 20 coordinate systems adopted by the Rocky Mountain Tribal Transportation Authority.  These CRS's are basically high precision CRS's for use on Indian Reservations in Montana and Wyoming.  See: www.neciusa.com/documents/RMTCRS_Manual.pdf for a complete description.
This particular submission contains two test points manually transcribed from the referenced document. Per the document (on or about page 34), I requested access to an electronic version of the rather substantial test point database printed in Table 4.2 in the referenced document.  As of this submission, I have not received a response to that request.